+++ src/components/calcite-loader/calcite-loader.stories.js
@@ -1,45 +1,79 @@
-import { storiesOf } from '@storybook/html';
-import { withKnobs, text, number, boolean, color } from '@storybook/addon-knobs'
-import { darkBackground, parseReadme } from '../../../.storybook/helpers';
-import readme from './readme.md';
+import { storiesOf } from "@storybook/html";
+import {
+  withKnobs,
+  text,
+  number,
+  boolean,
+  color,
+  select,
+} from "@storybook/addon-knobs";
+import { darkBackground, parseReadme } from "../../../.storybook/helpers";
+import readme from "./readme.md";
 const notes = parseReadme(readme);
 
-storiesOf('Loader', module)
+storiesOf("Loader", module)
   .addDecorator(withKnobs)
-  .add('Indeterminate', () => `
+  .add(
+    "Simple",
+    () => `
     <calcite-loader
-      is-active="${boolean("is-active", true)}"
-      text="${text("text", "")}"
-    />
-  `, { notes })
-  .add('Determinate', () => `
-    <calcite-loader
-      is-active
-      type="determinate"
+      active
+      type="${select(
+        "type",
+        ["determinate", "indeterminate"],
+        "indeterminate"
+      )}"
+      scale="${select("scale", ["s", "m", "l"], "m")}"
       no-padding="${boolean("no-padding", false)}"
-      value="${number('value', 0, {range: true, min: 0, max: 100, step: 1})}"
+      value="${number("value", 0, { range: true, min: 0, max: 100, step: 1 })}"
     />
-  `, { notes })
-  .add('Inline', () => `
+  `,
+    { notes }
+  )
+  .add(
+    "Inline",
+    () => `
+  <div style="display: inline-flex;align-items: center;justify-content: center;width: 100%;">
     <calcite-loader
+      scale="${select("scale", ["s", "m", "l"], "m")}"
       inline
-      is-active
-    /></calcite-loader>Next to some text
-  `, { notes })
-  .add('Dark mode', () => `
+      active
+    /></calcite-loader><span style="margin:0 10px">Next to some text</span>
+    </div>
+  `,
+    { notes }
+  )
+  .add(
+    "Dark mode",
+    () => `
     <calcite-loader
+      type="${select(
+        "type",
+        ["determinate", "indeterminate"],
+        "indeterminate"
+      )}"
+      scale="${select("scale", ["s", "m", "l"], "m")}"
       no-padding="${boolean("no-padding", false)}"
-      value="${number('value', 0, {range: true, min: 0, max: 100, step: 1})}"
-      is-active
+      value="${number("value", 0, { range: true, min: 0, max: 100, step: 1 })}"
+      active
       theme="dark" />
-  `, { notes, backgrounds: darkBackground })
-  .add('Custom theme', () => `
+  `,
+    { notes, backgrounds: darkBackground }
+  )
+  .add(
+    "Custom theme",
+    () => `
     <calcite-loader
-
+    type="${select("type", ["determinate", "indeterminate"], "indeterminate")}"
+    scale="${select("scale", ["s", "m", "l"], "m")}"
+    no-padding="${boolean("no-padding", false)}"
+    value="${number("value", 0, { range: true, min: 0, max: 100, step: 1 })}"
       style="
-        --calcite-loader-spot-light: ${color('spot-light', '#50ba5f')};
-        --calcite-loader-spot-dark: ${color('spot-dark', '#1a6324')};
-        --calcite-loader-spot: ${color('loader-spot', '#338033')};"
-      is-active
+        --calcite-loader-spot-light: ${color("spot-light", "#50ba5f")};
+        --calcite-loader-spot-dark: ${color("spot-dark", "#1a6324")};
+        --calcite-loader-spot: ${color("loader-spot", "#338033")};"
+      active
     />
-  `, { notes })
+  `,
+    { notes }
+  );